chore: [#947] Separate environment properties from utils\loader.py to  schemas/properties/env.py and move schemas/data.py to schemas/properties/data.py
