+++ transform/schemas/properties/data.py
@@ -1,0 +1,54 @@
+"""Define properties name that are used multiple times"""
+# TODO INPUT, OUTPUT, BOTH
+
+AUTHOR = "author"
+AUTHOR_NAMES = "author_names"
+AUTHOR_PIDS = "author_pids"
+BEST_ACCESS_RIGHT = "best_access_right"
+CONTENT_TYPE = "content_type"
+CONTEXT = "context"
+COUNTRY = "country"
+DATA_SOURCE = "datasource_pids"
+DOCUMENT_TYPE = "document_type"
+DOI = "doi"
+DOWNLOADS = "usage_counts_downloads"
+EOSC_IF = "eosc_if"
+EXPORTATION = "exportation"
+FOS = "fos"
+FUNDER = "funder"
+GEO_AV = "geographical_availabilities"
+ID = "id"
+INSTANCE = "instance"
+KEYWORDS = "keywords"
+LANGUAGE = "language"
+LVL_OF_EXPERTISE = "level_of_expertise"
+OPEN_ACCESS = "open_access"
+PERSIST_ID_SYS = "persistent_identity_systems"
+PERSIST_ID_SYS_ENTITY_TYPE = "persistent_identity_systems_entity_type"
+PERSIST_ID_SYS_ENTITY_TYPE_SCHEMES = "persistent_identity_systems_entity_type_schemes"
+PID = "pid"
+PIDS = "pids"
+POPULARITY = "popularity"
+PROJECTS = "projects"
+PUBLICATION_DATE = "publication_date"
+PUBLISHER = "publisher"
+RELATIONS = "relations"
+RELATIONS_LONG = "relations_long"
+RESEARCH_COMMUNITY = "research_community"
+RESOURCE_TYPE = "resource_type"
+SCIENTIFIC_DOMAINS = "scientific_domains"
+SDG = "sdg"
+SUBJECT = "subject"
+TAG_LIST = "tag_list"
+TARGET_GROUP = "target_group"
+TITLE = "title"
+TYPE = "type"
+UNIFIED_CATEGORIES = "unified_categories"
+URL = "url"
+URI = "uri"
+VIEWS = "usage_counts_views"
+
+# Text general
+AUTHOR_NAMES_TG = "author_names_tg"
+KEYWORDS_TG = "keywords_tg"
+TAG_LIST_TG = "tag_list_tg"
