--- transform/schemas/properties_name.py
@@ -1,52 +1,0 @@
-"""Define properties name that are used multiple times"""
-
-AUTHOR_NAMES = "author_names"
-AUTHOR_PIDS = "author_pids"
-AUTHOR = "author"
-BEST_ACCESS_RIGHT = "best_access_right"
-COUNTRY = "country"
-DOCUMENT_TYPE = "document_type"
-EXPORTATION = "exportation"
-FUNDER = "funder"
-ID = "id"
-KEYWORDS = "keywords"
-LANGUAGE = "language"
-OPEN_ACCESS = "open_access"
-PUBLISHER = "publisher"
-RESEARCH_COMMUNITY = "research_community"
-FOS = "fos"
-SDG = "sdg"
-SUBJECT = "subject"
-TITLE = "title"
-TYPE = "type"
-URL = "url"
-DOI = "doi"
-URI = "uri"
-UNIFIED_CATEGORIES = "unified_categories"
-DOWNLOADS = "usage_counts_downloads"
-VIEWS = "usage_counts_views"
-TAG_LIST = "tag_list"
-PROJECTS = "projects"
-INSTANCE = "instance"
-CONTEXT = "context"
-PID = "pid"
-RELATIONS = "relations"
-RELATIONS_LONG = "relations_long"
-PERSIST_ID_SYS = "persistent_identity_systems"
-PERSIST_ID_SYS_ENTITY_TYPE = "persistent_identity_systems_entity_type"
-PERSIST_ID_SYS_ENTITY_TYPE_SCHEMES = "persistent_identity_systems_entity_type_schemes"
-CONTENT_TYPE = "content_type"
-LVL_OF_EXPERTISE = "level_of_expertise"
-GEO_AV = "geographical_availabilities"
-RESOURCE_TYPE = "resource_type"
-SCIENTIFIC_DOMAINS = "scientific_domains"
-TARGET_GROUP = "target_group"
-PUBLICATION_DATE = "publication_date"
-EOSC_IF = "eosc_if"
-PIDS = "pids"
-POPULARITY = "popularity"
-
-# Text general
-AUTHOR_NAMES_TG = "author_names_tg"
-KEYWORDS_TG = "keywords_tg"
-TAG_LIST_TG = "tag_list_tg"
