+++ transform/transformations/common.py
@@ -5,46 +5,48 @@ from collections import defaultdict
 from itertools import chain
 from logging import getLogger
 import json
+
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import col, to_date, when, lit
 from pyspark.sql.utils import AnalysisException
-from schemas.properties_name import (
+
+from schemas.properties.data import (
     AUTHOR,
     AUTHOR_NAMES,
+    AUTHOR_NAMES_TG,
     AUTHOR_PIDS,
     BEST_ACCESS_RIGHT,
     COUNTRY,
-    DOCUMENT_TYPE,
-    FUNDER,
-    EXPORTATION,
-    OPEN_ACCESS,
-    RESEARCH_COMMUNITY,
-    TYPE,
-    URL,
-    DOI,
-    PUBLISHER,
-    FOS,
-    SDG,
-    UNIFIED_CATEGORIES,
-    DOWNLOADS,
-    VIEWS,
-    SUBJECT,
-    LANGUAGE,
-    TAG_LIST_TG,
-    TAG_LIST,
-    KEYWORDS_TG,
-    KEYWORDS,
-    AUTHOR_NAMES_TG,
-    PROJECTS,
-    INSTANCE,
     CONTEXT,
+    DOCUMENT_TYPE,
+    DOI,
+    DOWNLOADS,
+    EOSC_IF,
+    EXPORTATION,
+    FOS,
+    FUNDER,
+    INSTANCE,
+    KEYWORDS,
+    KEYWORDS_TG,
+    LANGUAGE,
+    OPEN_ACCESS,
     PID,
+    PIDS,
+    POPULARITY,
+    PROJECTS,
+    PUBLISHER,
+    RESEARCH_COMMUNITY,
     RELATIONS,
     RELATIONS_LONG,
-    EOSC_IF,
-    PIDS,
     SCIENTIFIC_DOMAINS,
-    POPULARITY,
+    SDG,
+    SUBJECT,
+    TAG_LIST,
+    TAG_LIST_TG,
+    TYPE,
+    UNIFIED_CATEGORIES,
+    URL,
+    VIEWS,
 )
 from schemas.mappings import (
     OPEN_ACCESS_,
