+++ transform/transformers/base/oag.py
@@ -13,7 +13,7 @@ from pyspark.sql.types import (
 from transformations.common import *
 from transformers.base.base import BaseTransformer
 from utils.utils import sort_schema
-from schemas.properties_name import *
+from schemas.properties.data import *
 
 
 class OagBaseTransformer(BaseTransformer):
