+++ transform/utils/loader.py
@@ -7,69 +7,7 @@ from dotenv import load_dotenv
 from pyspark.sql import SparkSession
 from conf.s3 import connect_to_s3
 from schemas.output import *
-
-
-DATASET = "DATASET"
-PUBLICATION = "PUBLICATION"
-SOFTWARE = "SOFTWARE"
-OTHER_RP = "OTHER_RP"
-TRAINING = "TRAINING"
-SERVICE = "SERVICE"
-DATASOURCE = "DATASOURCE"
-PROVIDER = "PROVIDER"
-GUIDELINE = "GUIDELINE"
-OFFER = "OFFER"
-BUNDLE = "BUNDLE"
-
-DATASET_PATH = "DATASET_PATH"
-PUBLICATION_PATH = "PUBLICATION_PATH"
-SOFTWARE_PATH = "SOFTWARE_PATH"
-OTHER_RP_PATH = "OTHER_RP_PATH"
-SERVICE_PATH = "SERVICE_PATH"
-DATASOURCE_PATH = "DATASOURCE_PATH"
-PROVIDER_PATH = "PROVIDER_PATH"
-OFFER_PATH = "OFFER_PATH"
-BUNDLE_PATH = "BUNDLE_PATH"
-
-GUIDELINE_ADDRESS = "GUIDELINE_ADDRESS"
-TRAINING_ADDRESS = "TRAINING_ADDRESS"
-
-INPUT_FORMAT = "INPUT_FORMAT"
-OUTPUT_FORMAT = "OUTPUT_FORMAT"
-OUTPUT_PATH = "OUTPUT_PATH"
-OUTPUT_SCHEMA = "OUTPUT_SCHEMA"
-
-SEND_TO_SOLR = "SEND_TO_SOLR"
-SOLR_ADDRESS = "SOLR_ADDRESS"
-SOLR_PORT = "SOLR_PORT"
-SOLR_DATASET_COLS = "SOLR_DATASET_COLS"
-SOLR_PUBLICATION_COLS = "SOLR_PUBLICATION_COLS"
-SOLR_SOFTWARE_COLS = "SOLR_SOFTWARE_COLS"
-SOLR_OTHER_RP_COLS = "SOLR_OTHER_RP_COLS"
-SOLR_TRAINING_COLS = "SOLR_TRAINING_COLS"
-SOLR_SERVICE_COLS = "SOLR_SERVICE_COLS"
-SOLR_DATASOURCE_COLS = "SOLR_DATASOURCE_COLS"
-SOLR_PROVIDER_COLS = "SOLR_PROVIDER_COLS"
-SOLR_GUIDELINE_COLS = "SOLR_GUIDELINE_COLS"
-SOLR_OFFER_COLS = "SOLR_OFFER_COLS"
-SOLR_BUNDLE_COLS = "SOLR_BUNDLE_COLS"
-
-SEND_TO_S3 = "SEND_TO_S3"
-S3_ACCESS_KEY = "S3_ACCESS_KEY"
-S3_SECRET_KEY = "S3_SECRET_KEY"
-S3_ENDPOINT = "S3_ENDPOINT"
-S3_BUCKET = "S3_BUCKET"
-S3_CLIENT = "S3_CLIENT"
-S3_DUMP_NAME = "S3_DUMP_NAME"
-
-CREATE_LOCAL_DUMP = "CREATE_LOCAL_DUMP"
-LOCAL_DUMP_PATH = "LOCAL_DUMP_PATH"
-
-ALL_COLLECTION = "ALL_COLLECTION"
-SEPARATE_COLLECTION = "SEPARATE_COLLECTION"
-SOLR_COL_NAMES = "SOLR_COL_NAMES"
-PATH = "PATH"
-ADDRESS = "ADDRESS"
+from schemas.properties.env import *
 
 solr_all_col_mapping = {
     SERVICE: SOLR_SERVICE_COLS,
